Create UPDATE MNSURE MAXIS WEB-I REPORT.vbs
This script requires you to first clean up the data received from Web-I to eliminate a few fields. Then, it will run on MAXIS and case note the PMI, MNsure ID and Integrated Case Number of the individuals on the MAXIS case. It will also collect the worker's X1 number, the worker's name and the supervisor's name.